+++ EasySequence/Classes/Sequences/EZSOrderedSet.m
@@ -91,13 +91,10 @@
 }
 
 - (void)removeObject:(id)anObject {
-    NSUInteger index = ({
-        EZS_SCOPELOCK(_orderedSetLock);
-        [EZS_Sequence(_orderedSet) firstIndexWhere:EZS_isEqual(anObject)];
-    });
-    
+    EZS_SCOPELOCK(_orderedSetLock);
+    NSUInteger index = [EZS_Sequence(_orderedSet) firstIndexWhere:EZS_isEqual(anObject)];
     if (index != NSNotFound) {
-        [self removeObjectAtIndex:index];
+        [_orderedSet removeObjectAtIndex:index];
     }
 }